move, simplify, add test
